minimum working parts of create report step 1
