+++ front-end/src/app/reports/create-workflow/create-report-step1.component.html
@@ -5,59 +5,37 @@
   <form id="form" [ngClass]="{ 'ng-submitted': formSubmitted }">
     <div class="grid">
       <div class="col-12">
-        <h3>Type of report</h3>
+        <h3>Report type</h3>
       </div>
-    </div>
-    <div class="grid">
-      <div class="col-6">
+      <div class="col-12">
         <div class="field">
-          <label for="type">(CHOOSE ONE)</label>
-          <div *ngFor="let reportTypeCategory of reportTypeCategories" class="p-field-checkbox">
-            <p-radioButton
-              [inputId]="reportTypeCategory"
-              name="reportTypeCategory"
-              [value]="reportTypeCategory"
-              [(ngModel)]="selectedReportTypeCategory"
-            ></p-radioButton>
-            <label [for]="reportTypeCategory">{{ reportTypeCategory | uppercase }}</label>
-          </div>
+          <label for="report_type_category">TIME PERIOD</label>
+          <p-selectButton
+            [options]="getReportTypeCategories()"
+            [(ngModel)]="selectedReportTypeCategory"
+            (ngModelChange)="onReportTypeCategoryChange()"
+            [ngModelOptions]="{ standalone: true }"
+          ></p-selectButton>
+          <app-error-messages fieldName="report_type_category" [formSubmitted]="formSubmitted"></app-error-messages>
         </div>
       </div>
-      <div class="col-6"></div>
-    </div>
-    <div class="grid">
-      <div class="col-12">
-        <hr />
-      </div>
-    </div>
-    <div class="grid">
-      <div class="col-12">
-        <h3>{{ selectedReportTypeCategory }}</h3>
-      </div>
-      <div class="col-12">
+      <div class="col-6">
         <div class="field grid">
-          <div
-            *ngFor="let reportType of getReportTypes()"
-            class="p-field-checkbox col-{{ Math.floor(Math.max(2, 12 / getReportTypes().length)) }}"
-          >
+          <label for="type">(CHOOSE ONE)</label>
+          <div *ngFor="let reportCode of getReportCodes()" class="p-field-checkbox col-12">
             <p-radioButton
-              [inputId]="reportType"
-              name="reportType"
-              [value]="reportType"
-              [(ngModel)]="selectedReportType"
+              [inputId]="reportCode"
+              name="reportCode"
+              [value]="reportCode"
+              [(ngModel)]="selectedReportCode"
+              [ngModelOptions]="{ standalone: true }"
             ></p-radioButton>
-            <label [for]="reportType">{{ reportType | label: f3xReportCodeLabels }}</label>
+            <label [for]="reportCode">{{ reportCode | label: f3xReportCodeLabels }}</label>
           </div>
         </div>
       </div>
-      {{
-        [
-          F3xReportTypeCategories.TWELVE_DAY_PRE_ELECTION_REPORT,
-          F3xReportTypeCategories.THIRTY_DAY_POST_ELECTION_REPORT
-        ].includes(selectedReportTypeCategory)
-      }}
       <ng-container *ngIf="isElectionReport()">
-        <div class="grid">
+        <div class="col-6">
           <div class="col-6">
             <div class="field">
               <label for="election_date">ELECTION ON</label>
