+++ front-end/src/app/reports/report-list/report-list.component.html
@@ -1,1 +1,1 @@
-REPORT LIST <app-create-report-step-1> </app-create-report-step-1>
+REPORT LIST <app-create-report-step1> </app-create-report-step1>
