+++ front-end/src/app/reports/reports.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ReportsRoutingModule } from './reports-routing.module';
 import { ReportListComponent } from './report-list/report-list.component';
-import { CreateReportStep1 } from './create-workflow/create-report-step-1.component';
+import { CreateReportStep1 } from './create-workflow/create-report-step1.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 // PrimeNG
@@ -11,6 +11,7 @@ import { RadioButtonModule } from 'primeng/radiobutton';
 import { CalendarModule } from 'primeng/calendar';
 import { SharedModule } from 'app/shared/shared.module';
 import { DropdownModule } from 'primeng/dropdown';
+import { SelectButtonModule } from 'primeng/selectbutton';
 
 @NgModule({
   declarations: [ReportListComponent, CreateReportStep1],
@@ -23,6 +24,7 @@ import { DropdownModule } from 'primeng/dropdown';
     DropdownModule,
     ReportsRoutingModule,
     RadioButtonModule,
+    SelectButtonModule,
     SharedModule,
   ],
 })
